+++ src/pyhf/tensor/common.py
@@ -61,3 +61,20 @@ class _TensorViewer(object):
             tensorlib.einsum('j...->...j', tensorlib.gather(data, idx))
             for idx in indices
         ]
+
+
+def _tensorviewer_from_slices(slices, names, batch_size):
+    target_slices = []
+    start = 0
+    for sl in slices:
+        stop = start + (sl.stop - sl.start)
+        target_slices.append(slice(start, stop))
+        start = stop
+
+    db = default_backend
+    ranges = []
+    for sl in target_slices:
+        ranges.append(db.astensor(range(sl.start, sl.stop)))
+    if not ranges:
+        return None
+    return _TensorViewer(ranges, names=names, batch_size=batch_size)
